feat(core): Add option for on-demand outbox creation
Also, add method that processes an on-demand oubox. This
method will be called by the listener that is framework specific
and will process the event emitted when saving the outbox.

The reason that we use the event publisher and listener is because
we want to processes the outbox in a separate transaction however,
the library is not aware about transactionality. To overcome this,
we need to hand over control to the caller by letting them implement
a publisher and a listener.