feat: add locking in cleanup
Despite the fact that the store implementation could simply issue a
`DELETE FROM outbox_item WHERE delete_after < now`
it could also, fetch the to be deleted items and then try to delete
them, which when run concurrently (w/o locks) could end up with
errors. Thus, to be safe, we're introducing locks in cleanup, as well.

We could use the same locks provider we're using for monitor, but that
would serialize the monitor and cleanup runs, which is not required.

[VR-3403]