chore: mechanism to reset an `OutboxItemProcessor`
Implements a `reset` method on `OutboxItemProcessor` that stops any
further execution of the `Runnable` and resets the `OutboxItem` if
appropriate.

This encapsulates the logic for cleaning up the processing of an item
including resetting it to an appropriate state so that it will be
picked up on the next run. At the same time it should allow for the
removal of the `getItem` method which exposes the internal state and
implementation of `OutboxItemProcessor`.

[SW-587]