chore: use `OutboxItemProcessor#reset` for resetting items
Updates `TransactionalOutboxImpl` to utilize the newly introduced
`OutboxItemProcessor#reset` method for resetting outbox items when the
corresponding processor does not start. This accounts both for a failure
when submitting to the executor, as well as for the executor being
shutdown before the processor had the chance to run.

There is a known bug that causes the predicate applied to
`notExecutedRunnables` to not match any items. This means that the
corresponding code fragment never gets executed. However, this is
irrelevant to the current change and once the bug gets fixed, the
`stop` method will be properly called and reset the outbox item.

Finally, since it is no longer used, `OutboxItemProcessor#getItem` is
removed.

Reference:
https://github.com/bluegroundltd/transactional-outbox/discussions/24

[SW-587]