chore: drop ineffective concurrent modification guard
Removes the `isStopped` flag which was meant to provide some sort of
protection against concurrent modification of the outbox item state
through `run` and `reset`. However, since the methods are not
synchronized and may be (and are) called from different threads the
protection was rather weak and could lead to a false sense of safety.

Still, we do keep the check against the status of the outbox item. The
purpose of this is to allow for the method to be called 'freely'
without the need for the called to know the status of the outbox item
which would break the encapsulation and leak implementation details.

In the future, if needed, we could consider implementing a proper
synchronization mechanism.

[SW-587]