chore: introduce component to serve as a host for processing
Introduces `OutboxProcessingHost` which serves as a host inside which
an `OutboxItemProcessor` runs. This allows for separating the
responsibilities of processing the items and being scheduled on an
executor (i.e. implementing `Runnable`).

Due to the decoration process there is still the requirement for
implementing `Runnable` on `OutboxItemProcessor`, but this will be
removed in subsequent commits.

[SW-587]