feat: unify error handling during processing
Adapts `OutboxItemProcessor#run` so that all errors are implemented
and handled in a common pattern and their cause is propagated to the
caller through an exception. This is now allowed since the class no
longer implements `Runnable` and gives the caller a chance to
implement additional mitigating actions.

Since the current implementation (see `OutboxProcessingHost`) wraps the
processor in a `try/catch` block and resets it on all errors, this has
the (desired) side effect of resetting the state of the outbox item
when an appropriate handler is not provided. For the rest of the
errors, the 'reset' method has no effect since the outbox item status
will have already been updated to something other than `RUNNING`.

[SW-587]