refactor: introduce handler resolver in item processor
Updates `OutboxItemProcessor` so that it receives a handler resolver
function instead of an already pre-selected `OutboxHandler`.

This has the effect that the resolution of the proper outbox handler
is delayed until the item is processed. While this is not desirable
from a debugging perspective, it means that in (the rare) case of a
missing (or otherwise invalid) handler, only the relevant items will
be affected. The rest will be processed normally.

[SW-587]