refactor: introduce processing host builder
Introduces `OutboxProcessingHostBuilder` which is injected to
`TransactionalOutboxImpl` and is used to construct
`OutboxProcessingHost` instances.

The main purpose of this is to enhance/simplify testing since it allows
for mocking the instances that are submitted to the executor.
Accordingly, a lot of the unit tests have been adapted to make use of
this functionality.

[SW-587]