feat: introduce outbox processor for item groups
Introduces `OutboxGroupProcessor` that is responsible for encapsulating
the processing of an `OutboxItemGroup`. The implementation simply
instantiates an `OutboxItemProcessor` for each oh the items contained
in the group and executes them sequentially.

If any item fails to be processed, the flow is stopped with no further
action and the exception is propagated to the caller. It is the
responsibility of the caller to provide any mitigating actions if
required. (The current implementation calls `reset` which effectively
will reset the state of all items that are still `RUNNING`.)

Finally, `TransactionalOutboxImpl` is adjusted so that it builds an
`OutboxGroupProcessor` with a group containing a single item.

[SW-587]