test: improve dummy outbox handlers
Adds proper implementations for all methods in `DummyOutboxHandler`
to allow for a better testing experience. This brings it on par with
`DummyHandler` with the only meaningful difference being that the
latter implements a `sleep` operation when handling an item.
Accordingly, the latter is renamed to `DelayingOutboxHandler` to better
denote its usage.

[SW-587]