feat: throw exception when processing non-running outbox
Updates `OutboxItemProcessor` so that (as a first step in `run`) it
checks that the status of the `OutboxItem` is indeed `RUNNING`. If not,
an exception is thrown.

This should have always been the case, but is now even more important
due to the inclusion of failed items in the processing pipeline.

[SW-588]