feat: include failed items when setting up processing
Updates `TransactionalOutboxImpl` so that it includes `FAILED` items
when marking them for processing and eventually submitting them to the
executor.

**IMPORTANT**
This does not mean that `FAILED` items are fetched. This is abstracted
in the implementation of `OutboxStore#fetch` and is therefore the
responsibility of each application that utilizes the library.

[SW-588]