chore: Refactor `processInstantOutbox` to use monitor with hint
This is required in order to be able to execute all items that
potentially belong to the same group as the one being instantly
processed.

Since hint (outbox id) is passed through the filter, clients
can manipulate how the item is retrieved via `fetch`. To ensure
that we execute only this item in case of an erroneous fetch
implementation, we also filter it on runtime.

Also added `@Deprecated` since this method will eventually be
dropped in favor of directly using monitor with a hint.

[SW-581]