feat: enhance processing eligibility verification in processor
Updates `OutboxItemProcessor` to use the newly introduced
`markedForProcessing` property to verify that the item should indeed be
processed. This ensures that the check is consistent with the logic in
`TransactionalOutboxImpl`. The latter is protected by a lock and
therefore ensures that no items will concurrently be picked up by
multiple monitor cycles.

[SW-580]