fix: adjust item filtering to work with groups
Adjusts the single item filtering logic in `monitor` (i.e. when invoked
with a specified `id`) to work with groups.

The original implementation was filtering the items immediately after
they were fetched. This is problematic with regard to the grouping
functionality since it allows for potentially filtering out items that
would belong to the same group as the matched item.

Instead, we allow for the groups to be built and then apply the filter
in order to select only the group(s) that contain the requested item.
This has the additional advantage of remaining totally agnostic to the
way the grouping is performed.

[SW-580]