fix: make instant outbox items eligible for processing
Since `OutboxItemProcessor` has been updated to check processing
eligibility through `markedForProcessing`, it has been failing for
legacy instant outboxes that don't move through the `monitor` flow.

Accordingly, we need to update the building of the instant outboxes
to make sure that `prepareForProcessing` is also called for them. It
turns out that the basic difference between `makeScheduledOutboxItem`
and `makeInstantOutbox` is actually exactly that, i.e. the latter is
also 'preparing' the item for processing (albeit without 'marking'
them). Therefore, `TransactionalOutboxImpl#add` has been refactored to
always call `makeScheduledOutboxItem` and then conditionally also call
`updateForProcessing` on the returned item.

Finally, this means that `makeInstantOutbox` is no longer used and was
therefore removed. This has the added benefit of reducing the
dependencies of `OutboxItemFactory`.

[SW-580]