feat: support hints when inserting items
Introduces the ability for transactional outbox to supply hints to
`OutboxStore` when inserting items. The hints provide additional
context for the current operation and the transactional outbox
configuration.

In turn, the store can use the hints to provide a more appropriate /
performant implementation for the insertion operation.

The functionality is implemented as `OutboxStore#insert(item, hints)`,
i.e. an overload of the existing `OutboxStore#insert(item)` method. It
has a default implementation that simply calls the latter. Therefore,
if the implementing code does not wish to take advantage of the
feature, they can simply ignore it.

[SW-580]